+++ src/mcp_learning/mcp_client/__init__.py
@@ -1,6 +1,6 @@
 """Implement a basic MCP client."""
 
-from .configurations import Configurations as MCPClientConfigurations
 from .main import main as mcp_client_main
+from .utils import Configurations as MCPClientConfigurations
 
 __all__ = ["MCPClientConfigurations", "mcp_client_main"]
