+++ src/mcp_learning/mcp_client/llm.py
@@ -14,7 +14,7 @@ from openai.types.chat import (
     ChatCompletionToolParam,
 )
 
-from .configurations import (
+from .utils import (
     AzureOpenAIConfigurations,
     Configurations,
     HostedOpenAIConfigurations,
