+++ src/mcp_learning/mcp_client/utils/log.py
@@ -1,30 +1,13 @@
-"""Provide console formatting utilities for user interaction."""
+"""Configure logging."""
 
 import logging
 import logging.config
 import typing
 
-import rich
 import rich.logging
-import rich.pretty
 
 from .configurations import Configurations
-from .utils import MissingOptionalDependencyError, validate_optional_dependency_installation
-
-CONSOLE = rich.get_console()
-
-try:
-    validate_optional_dependency_installation("prompt-toolkit", import_name="prompt_toolkit")
-except MissingOptionalDependencyError:
-    ENHANCED_CLI_AVAILABLE = False
-else:
-    ENHANCED_CLI_AVAILABLE = True
-
-    import prompt_toolkit
-    import prompt_toolkit.key_binding
-    import prompt_toolkit.styles
-
-    SESSION = prompt_toolkit.PromptSession()
+from .console import CONSOLE
 
 
 def create_rich_handler(level: int | str = logging.NOTSET) -> rich.logging.RichHandler:
@@ -144,72 +127,3 @@ def initiate_logging(settings: Configurations) -> None:
 
     logging.config.dictConfig(logging_configurations)
     logging.captureWarnings(True)
-
-
-async def user_prompt() -> str:
-    """Prompt the user for input with a custom format.
-
-    Returns
-    -------
-    str
-        user provided input
-    """
-    if not ENHANCED_CLI_AVAILABLE:
-        return CONSOLE.input(prompt="\n[bold blue][You][/bold blue] ")
-
-    key_bindings = prompt_toolkit.key_binding.KeyBindings()
-
-    @key_bindings.add("enter")
-    def submit_input(event: prompt_toolkit.key_binding.KeyPressEvent) -> None:
-        """Submit the input on Enter key press.
-
-        Parameters
-        ----------
-        event : prompt_toolkit.key_binding.KeyPressEvent
-            key press event
-        """
-        buffer = event.current_buffer
-
-        if (
-            event.current_buffer.document.is_cursor_at_the_end
-            or event.current_buffer.document.is_cursor_at_the_end_of_line
-        ):
-            buffer.validate_and_handle()
-        else:
-            buffer.insert_text("\n")
-
-    style = prompt_toolkit.styles.Style.from_dict({"prompt": "bold blue"})
-
-    prompt = await SESSION.prompt_async(
-        [("class:prompt", "\n[You] ")], key_bindings=key_bindings, style=style, multiline=True
-    )
-
-    return prompt
-
-
-def bot_response(message: typing.Any) -> None:  # noqa: ANN401
-    """Print the bot's response in a formatted way.
-
-    Parameters
-    ----------
-    message : typing.Any
-        message to print
-    """
-    if not isinstance(message, str):
-        message = rich.pretty.Pretty(message)
-
-    CONSOLE.print("[bold magenta][Bot][/bold magenta]", message)
-
-
-async def llm_response(token_stream: typing.AsyncIterable[str]) -> None:
-    """Print the LLM response in a formatted way.
-
-    Parameters
-    ----------
-    token_stream : typing.AsyncIterable[str]
-        stream of message to print
-    """
-    CONSOLE.print("[bold green][LLM][/bold green] ", end="")
-
-    async for token in token_stream:
-        CONSOLE.print(token, end="")
